+++ docs/package.json
@@ -26,7 +26,6 @@
     "moti": "^0.18.0",
     "next": "^12.1.0",
     "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "solito": "latest",
     "zeego": "latest"
