--- examples/expo/(root).tsx
@@ -1,26 +1,0 @@
-import { Tabs } from 'expo-router'
-
-export default function Root() {
-  return (
-    <Tabs initialRouteName="index">
-      <Tabs.Screen
-        name="index"
-        options={{
-          headerShown: false,
-        }}
-      />
-      <Tabs.Screen
-        name="twitter"
-        options={{
-          headerShown: false,
-        }}
-      />
-      <Tabs.Screen
-        name="tweet"
-        options={{
-          headerShown: false,
-        }}
-      />
-    </Tabs>
-  )
-}
