+++ examples/expo/babel.config.js
@@ -17,8 +17,6 @@ module.exports = function (api) {
           },
         },
       ],
-      // NOTE: `expo-router/babel` is a temporary extension to `babel-preset-expo`.
-      require.resolve('expo-router/babel'),
       'react-native-reanimated/plugin',
     ],
   }
