+++ examples/expo/package.json
@@ -14,26 +14,26 @@
   "dependencies": {
     "@react-native-menu/menu": "^0.7.2",
     "add": "^2.0.6",
-    "expo": "^46.0.0",
-    "expo-router": "^0.0.33",
-    "expo-splash-screen": "~0.16.2",
-    "expo-status-bar": "~1.4.0",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
-    "react-native": "0.69.6",
-    "react-native-gesture-handler": "~2.5.0",
+    "expo": "~51.0.28",
+    "expo-linking": "^6.3.1",
+    "expo-router": "~3.5.23",
+    "expo-splash-screen": "~0.27.5",
+    "expo-status-bar": "~1.12.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-native": "0.74.5",
+    "react-native-gesture-handler": "~2.16.1",
     "react-native-ios-context-menu": "2.3.2",
-    "react-native-reanimated": "~2.9.1",
-    "react-native-safe-area-context": "4.3.1",
-    "react-native-screens": "~3.15.0",
-    "react-native-web": "~0.18.7"
+    "react-native-reanimated": "~3.10.1",
+    "react-native-safe-area-context": "4.10.5",
+    "react-native-screens": "3.31.1",
+    "react-native-web": "~0.19.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.6",
+    "@babel/core": "^7.24.0",
     "@babel/runtime": "^7.9.6",
     "babel-loader": "^8.2.3",
     "babel-plugin-module-resolver": "^4.0.0",
-    "babel-preset-expo": "~9.2.0",
-    "expo-cli": "latest"
+    "babel-preset-expo": "~11.0.0"
   }
 }
