+++ package.json
@@ -51,11 +51,9 @@
     "lerna": "^3.22.1",
     "pod-install": "^0.1.0",
     "prettier": "^2.0.5",
-    "react": "17.0.1",
-    "react-native": "0.64.2",
     "react-native-builder-bob": "^0.18.0",
     "release-it": "^14.2.2",
-    "typescript": "^5"
+    "typescript": "5.6.2"
   },
   "peerDependencies": {
     "react": "*",
@@ -117,11 +115,7 @@
     ]
   },
   "resolutions": {
-    "react-native": "0.69.6",
-    "expo-modules-autolinking": "~0.10.1",
-    "@expo/config-plugins": "^5.0.0",
-    "@expo/prebuild-config": "^5.0.1",
-    "react": "18.0.0",
+    "react": "18.2.0",
     "metro": "0.72.3"
   },
   "packageManager": "yarn@4.4.0"
