+++ packages/zeego/src/dropdown-menu/index.tsx
@@ -1,2 +1,2 @@
 export * from './dropdown-menu'
-export { create, menuify } from '../menu'
+export { create } from '../menu'
